Update instruction list and add user agent header.
Since this was written 5 years ago, the Navy public website now has a
firewall configured to try to block bots.  The first check is surmounted
by specifying a popular User-Agent (in this case, Googlebot).

There is a second check using a simple CAPTCHA. Not simple for me to
automate so this is still broken, but maybe a fancy enough problem for
an AI? :)